+++ src/Foundation/Connect/code/HCA.Foundation.Connect.csproj
@@ -575,13 +575,13 @@
     <Compile Include="Context\Storefront\StorefrontContext.cs" />
     <Compile Include="Loaders\CommerceTypeLoader.cs" />
     <Compile Include="Loaders\ICommerceTypeLoader.cs" />
-    <Compile Include="Managers\Account\AccountManagerV2.cs" />
-    <Compile Include="Managers\Account\IAccountManagerV2.cs" />
+    <Compile Include="Managers\Account\AccountManager.cs" />
+    <Compile Include="Managers\Account\IAccountManager.cs" />
     <Compile Include="Managers\Analytics\AnalyticsManager.cs" />
     <Compile Include="Managers\Analytics\IAnalyticsManager.cs" />
     <Compile Include="Managers\BaseManager.cs" />
-    <Compile Include="Managers\Cart\CartManagerV2.cs" />
-    <Compile Include="Managers\Cart\ICartManagerV2.cs" />
+    <Compile Include="Managers\Cart\CartManager.cs" />
+    <Compile Include="Managers\Cart\ICartManager.cs" />
     <Compile Include="Managers\Inventory\IInventoryManager.cs" />
     <Compile Include="Managers\Inventory\InventoryManager.cs" />
     <Compile Include="Managers\Order\OrderManager.cs" />
@@ -631,11 +631,6 @@
     <Compile Include="Models\Search\SortDirection.cs" />
     <Compile Include="Models\ShippingInfoArgument.cs" />
     <Compile Include="Models\Catalog\Variant.cs" />
-    <Compile Include="Managers\AccountManager.cs" />
-    <Compile Include="Managers\CartManager.cs" />
-    <Compile Include="Managers\Contracts\ICartManager.cs" />
-    <Compile Include="Managers\Contracts\IAccountManager.cs" />
-    <Compile Include="Managers\ManagerResponse.cs" />
     <Compile Include="Models\CartLineArgument.cs" />
     <Compile Include="Models\Search\SearchResults.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
